+++ backend/SC4SNMP_UI_backend/apply_changes/routes.py
@@ -1,7 +1,9 @@
-from flask import Blueprint, jsonify
+from flask import Blueprint, jsonify, current_app
 from flask_cors import cross_origin
 from SC4SNMP_UI_backend.apply_changes.apply_changes import ApplyChanges
+from SC4SNMP_UI_backend.apply_changes.handling_chain import EmptyValuesFileException, YamlParserException
 import os
+import traceback
 
 apply_changes_blueprint = Blueprint('common_blueprint', __name__)
 JOB_CREATION_RETRIES = int(os.getenv("JOB_CREATION_RETRIES", 10))
@@ -20,3 +22,14 @@ def apply_changes():
         message = f"Configuration will be updated in approximately {job_delay} seconds."
     result = jsonify({"message": message})
     return result, 200
+
+@apply_changes_blueprint.errorhandler(Exception)
+@cross_origin()
+def handle_exception(e):
+    current_app.logger.error(traceback.format_exc())
+    if isinstance(e, (EmptyValuesFileException, YamlParserException)):
+        result = jsonify({"message": e.message})
+        return result, 400
+
+    result = jsonify({"message": "Undentified error. Check logs."})
+    return result, 400